Rename SuccessorNodes to SuccessorsFunc
The original name is hold over from early development and probably shouldn't have lasted as long as it did. The type is a function which returns successors, so `SuccessorNodes` is more suitable to describe the *return* value of the function. For the function itself we can be more explicit and include `Func` in the name. That also means we don't have to keep adding the word "function" when referring to the type.