Move usage documentation from source file to README
A bunch of fix up inspired by seeing the package on Hex for the first. At this stage it looks like it will be more usable if the documentation is in the README, rather than at the top of the source file. The docs at the top of the source file are a good spot to put navigation aids to the rest of the documentation, but otherwise the most accessible spot is the README, which appears as the front page for the Hex package documentation. It therefore doesn't make much sense to add a link to the Hex package documentation in the README for people coming from GitHub. I suppose I can use the GitHub description for that instead.

Also fixed up a few markdown tags, and removed/integrated some draft text.